Pawn_Solo/requirements.txt
python-chess
numpy
tensorflow

#a sense of pity